+++ src/controllers/console/client/base/Oauth2BaseClientAction.php
@@ -5,6 +5,7 @@ namespace rhertogh\Yii2Oauth2Server\controllers\console\client\base;
 use rhertogh\Yii2Oauth2Server\controllers\console\Oauth2ClientController;
 use rhertogh\Yii2Oauth2Server\models\Oauth2Client;
 use yii\base\Action;
+use yii\base\InvalidConfigException;
 use yii\console\Exception;
 
 /**
@@ -12,21 +13,16 @@ use yii\console\Exception;
  */
 class Oauth2BaseClientAction extends Action
 {
+    /**
+     * Find a client by its id or identifier.
+     *
+     * @throws Exception
+     * @throws InvalidConfigException
+     */
     protected function findByIdOrIdentifier($id)
     {
         $module = $this->controller->module;
-
-        try {
-            /** @var Oauth2Client $client */
-            $client = $module->getClientRepository()->findModelByPk($id);
-        } catch (\Exception $e) {
-            // Silently ignore.
-        }
-
-        if (empty($client)) {
-            // try to find by `identifier`.
-            $client = $module->getClientRepository()->findModelByIdentifier($id);
-        }
+        $client = $module->getClientRepository()->findModelByPkOrIdentifier($id);
 
         if (empty($client)) {
             throw new Exception('No client with id or identifier "' . $id . '" found.' . PHP_EOL);